% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{airport_distance}
\alias{airport_distance}
\title{Calculate great circle distance between two airports. Distances are calculated using the
Haversine formula which assumes a spherical earth. Distances are returned in kilometres}
\usage{
airport_distance(airport1, airport2)
}
\arguments{
\item{airport1}{Takes a three-letter IATA code corresponding to an airport}

\item{airport2}{as above}
}
\value{
Great circle distance in kilometres between the two airports
}
\description{
Calculate great circle distance between two airports. Distances are calculated using the
Haversine formula which assumes a spherical earth. Distances are returned in kilometres
}
\examples{
airport_distance("YVR","YYZ")
}
