% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airport_lookups.R
\name{airport_location}
\alias{airport_location}
\title{Return airport location in lon/lat given an input IATA code, ICAO code, or airport name.}
\usage{
airport_location(input, input_type)
}
\arguments{
\item{input}{An airport name, IATA code, or ICAO code}

\item{input_type}{One of "name", "IATA", or "ICAO"}
}
\value{
List of longitude and latitude coordinates
}
\description{
Return airport location in lon/lat given an input IATA code, ICAO code, or airport name.
}
\examples{
airport_location("YVR","IATA")
#' airport_location("Vancouver International Airport","name")
}
