% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airport_lookups.R
\name{airport_lookup}
\alias{airport_lookup}
\title{Return city name, airport name, IATA code, or IACO code given an input IATA code, ICAO code, or airport name.}
\usage{
airport_lookup(input, input_type, output_type)
}
\arguments{
\item{input}{An airport name, IATA code, or ICAO code}

\item{input_type}{One of "name", "IATA", or "ICAO"}

\item{output_type}{One of "name", "city", "IATA", or "ICAO"}
}
\value{
The appropriate city, airport name, IATA code, or ICAO code for that airport
}
\description{
Return city name, airport name, IATA code, or IACO code given an input IATA code, ICAO code, or airport name.
}
\examples{
airport_lookup("Vancouver International Airport","name","IATA")
airport_lookup("YVR","IATA","city")

Produces a list of similar named airports
airport_lookup("Vancoover","IATA","city")
}
