% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airportr.R
\docType{package}
\name{airportr-package}
\alias{airportr}
\alias{airportr-package}
\title{airportr: Convenience Tools for Working with Airport Data}
\description{
This package uses open source airport data to provide tools to look up
information, translate names and codes into other formats, as well as
some basic calculation functions for measuring distances.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dshkol/airportr}
  \item \url{https://dshkol.github.io/airportr/}
  \item Report bugs at \url{https://github.com/dshkol/airportr/issues}
}

}
\author{
\strong{Maintainer}: Dmitry Shkolnik \email{shkolnikd@gmail.com}

}
\keyword{internal}
