% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airportr.R
\docType{package}
\name{airportr}
\alias{airportr}
\title{\code{airportr} package}
\description{
Package to work with airport data
}
\details{
See the README on \href{https://github.com/dshkol/airportr/blob/master/README.md}{GitHub}
}
