% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{airports_near}
\alias{airports_near}
\title{Lookup airports within specified distance of a specified airport.}
\usage{
airports_near(input, distance = 100)
}
\arguments{
\item{input}{An airport name, IATA code, or ICAO code. Input type will be guessed unless
explicitly defined}

\item{distance}{Distance boundary for nearest airport lookup in kilometres}
}
\value{
A tibble of airports that fall within the specified range of input airport
}
\description{
Lookup airports within specified distance of a specified airport.
}
\examples{
airports_near("YVR")

# Or with a user specified distance in kilometres
airports_near("YVR", distance = 200)
}
