% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airport_lookups.R
\name{airport_detail}
\alias{airport_detail}
\title{Lookup full airport details based of a standard airport input}
\usage{
airport_detail(input, input_type)
}
\arguments{
\item{input}{An airport name, IATA code, or ICAO code. Input type will be guessed unless
explicitly defined}

\item{input_type}{One of "name", "IATA", or "ICAO". Function will attempt to guess type if
not supplied}
}
\value{
A 1x14 tibble with airport details
}
\description{
Return all airport details given an input IATA code, ICAO code, or airport name.
}
\examples{
airport_detail("YVR")
airport_detail("London Heathrow Airport")
}
