% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airport_lookups.R
\name{city_airports}
\alias{city_airports}
\title{Return all airports matching a city input}
\usage{
city_airports(city, country)
}
\arguments{
\item{city}{A city name. If no exact match will attempt to prompt user with suggested alternatives}

\item{country}{(Optional) A country name}
}
\value{
A nx14 tibble with airport details where n is the number of airports serving that city
}
\description{
Return all airports matching a city input
}
\examples{
city_airports("Vancouver")
city_airports("London")
city_airports("London","Canada")
}
